% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{bisect}
\alias{bisect}
\title{Bisection search for a forecast target providing a given performance statistic value.}
\usage{
bisect(
  stock,
  sr,
  deviances = rec(stock) \%=\% 1,
  metrics,
  refpts,
  statistic,
  years,
  pyears = years,
  tune,
  prob,
  tol = 0.01,
  maxit = 15,
  verbose = TRUE
)
}
\description{
The plain bisection algorithm (Burden & Douglas, 1985) is employed here to
find the value of a given forecast target quantity (e.g. \code{fbar}) for which
a selected value of a performance statistic is obtained over a chosen period.
}
\examples{
data(ple4)
stock <- propagate(stf(ple4, end=2118), 100)
srr <- predictModel(model=rec~a*ssb*exp(-b*ssb), params=FLPar(a=5.20, b=1.65e-6))
# generate srr deviances
devs <- ar1rlnorm(rho=0.4, 2018:2118, iters=100, meanlog=0, sdlog=0.5)
# define fp05 statistic
statistic <- list(fp05=list(~yearMeans((sb/sblim) < 1), name="p.05",
  desc="ices p.05"))
# call bisect over 100 years, fp.05 calculated over last 50.
fp05fwd <- bisect(stock, sr=srr, deviances=devs, metrics=list(sb=ssb), 
  refpts=FLPar(sblim=150000), statistic=statistic, years=2018:2118,
  pyears=2069:2118, tune=list(fbar=c(0.1, 1)), prob=0.05)
}
\references{
{Burden, Richard L.; Faires, J. Douglas (1985), "2.1 The Bisection Algorithm", Numerical Analysis (3rd ed.), PWS Publishers, ISBN 0-87150-857-5}
}
