% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{boot_mse}
\alias{boot_mse}
\title{Estimation of 95\% bootstrap confidence intervals.}
\usage{
boot_mse(object, modelnr = NULL, rep = 1000, seed = 1)
}
\arguments{
\item{object}{an object created with the function \code{\link{model_search}}.}

\item{modelnr}{the number of the model in \code{object} that is to be bootstrapped.
If left unspecified, the  model with the lowest BIC will be selected.}

\item{rep}{the number of bootstrap replications.}

\item{seed}{a scalar to set the random seed.}
}
\value{
A list with tables and, if the variable \code{year} is specified, a list
with plots.
\item{tables}{table(s) with the number of observations, the population size estimate
with the 95\% bootstrap confidence interval. If covariates are present, the results are
given for the levels of (pairs of) the covariates. If the variable \code{year} is
specified, these results are presented for the distinct periods of time.}
\item{plots}{if the covariate \code{year} is present, plots of the population size
estimates and 95\% confidence intervals are produced that show the estimated trends
over the distinct periods of time.}
}
\description{
\code{boot_mse} performs a parametric bootstrap for a selected model,
and computes 95\% confidence intervals.
}
