% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{catchSSB.hcr}
\alias{catchSSB.hcr}
\title{A HCR to set catch based on SSB}
\usage{
catchSSB.hcr(stk, dtarget = 0.4, dlimit = 0.1, lambda = 1, MSY,
  ssb_lag = 1, ay, tracking)
}
\description{
A HCR to set catch based on SSB
}
