% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{cpue.hcr}
\alias{cpue.hcr}
\title{cpue.hcr}
\usage{
cpue.hcr(
  stk,
  ay,
  k1,
  k2,
  k3,
  k4,
  target = 1,
  dtaclow = 0.85,
  dtacupp = 1.15,
  tracking
)
}
\description{
cpue.hcr
}
\examples{
data(ple4)
}
