% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind.R
\name{cpue.ind}
\alias{cpue.ind}
\title{Computes CPUE-based Indicators of changes in Stock Abundance}
\usage{
cpue.ind(stk, idx, index = 1, nyears = 5, args, tracking)
}
\arguments{
\item{stk}{An object representing the stock returned by the \code{oem} modules, FLStock.}

\item{idx}{An FLIndices containing the indices rturned by the \code{oem} module.}

\item{index}{An integer or character, specifying the index to use from \code{idx}. Default is 1.}

\item{nyears}{An integer, the number of years to consider for the calculations. Default is 5.}

\item{args}{A list containing dimensionality arguments, passed on by mp().}

\item{tracking}{An FLQuant used for tracking indicators, intermediate values, and decisions during MP evaluation.}
}
\value{
A list containing 'stk', the input FLStock, 'ind, an FLQuants object
containing the computed index metrics, and the 'tracking' table.
}
\description{
This function computes four abundance indicators from one CPUE or biomass
index of abundance: the index itself, an average over a number of years, a
weighted mean over those same years and the slope of the trend over the same period.
}
\details{
The weighted average returned in the 'wmean' element is calculated over the last
\code{nyears}. The last year's weight is set as 50\%, and the remaining years share the
other 50\% proportionally. The 'slope' metric is computed on the log-transformed data.

Three elements are added to the \code{tracking} table:
\itemize{
\item mean.ind, with the index average
\item wmean.ind, with the index weighted average
\item slope.ind, with the index slope
}
}
\examples{
data(plesim)
# MP control with CPUE: catch ~ weighted 4-year mean CPUE
ctrl <- mpCtrl(est=mseCtrl(method=cpue.ind, args=list(index=2)),
  hcr=mseCtrl(method=hockeystick.hcr, args=list(metric="wmean",
  trigger=2000, output="catch", target=1.25e5)))

# Run the MP
 run <- mp(om, oem, control=ctrl, args=list(iy=2025, fy=2035))

# Plot results
plot(om, run)
}
