% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db}
\alias{db}
\title{Creates a single table of output from an FLmse object}
\usage{
db(object, ...)
}
\arguments{
\item{object}{Input object with MSE results}

\item{metrics}{List of metrics to be computed on relevant slots}
}
\value{
A \code{data.table} object
}
\description{
Creates a single table of output from an FLmse object
}
\examples{
data(p4om)
db(om)
}
