% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mseCtrl-class.R,
%   R/mpCtrl-class.R, R/FLo-class.R
\name{debug-mse}
\alias{debug-mse}
\alias{debug,mseCtrl,missing-method}
\alias{undebug,mseCtrl,missing-method}
\alias{undebug,mpCtrl,character-method}
\alias{undebug,mpCtrl,missing-method}
\alias{debug,FLo,ANY-method}
\alias{undebug,FLo,ANY-method}
\title{Debugging mse modules}
\usage{
debug(fun, text = "", condition = NULL, signature = NULL)

\S4method{debug}{mseCtrl,missing}(fun)

\S4method{undebug}{mseCtrl,missing}(fun)

\S4method{undebug}{mpCtrl,character}(fun, signature = NULL)

\S4method{undebug}{mpCtrl,missing}(fun)

\S4method{debug}{FLo,ANY}(fun)

\S4method{undebug}{FLo,ANY}(fun)
}
\arguments{
\item{fun}{Module or control object to debug.}

\item{text}{Name of module in mpCtrl.}

\item{condition}{Unused.}

\item{signature}{Name of module in mpCtrl.}
}
\value{
Both functions invisibly return NULL
}
\description{
Set and unset the debugging flag of a function inside the \emph{method} slot of
a mseCtrl object.
}
\details{
Modules in the mse control object contain the function to be called in the
\emph{method} slot. To debug and check the behaviour of an individual function,
the \emph{debug} method will start a browser session next time it is called.
Debugging functions requires the parallel flag to be set to FALSE, or that
no parallel backend is loaded.

Calling \emph{undebug} on an mpCtrl without specifying a module will check for
the debugging status of each of them, and undebug if TRUE.

For objects of classes \emph{FLom} and \emph{FLombf}, \emph{debug} and \emph{undebug} will set
and unset the debugging flag on the function stored in the \emph{projection} slot.
}
\seealso{
\code{\link{debug}}
}
\author{
Iago Mosqueira (WMR)
}
\keyword{methods}
