% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{extractPerformance}
\alias{extractPerformance}
\title{Extracts performance time series for an MP including the corresponding historical OM}
\usage{
extractPerformance(dat, mp)
}
\description{
Extracts performance time series for an MP including the corresponding historical OM
}
