% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLmse-class.R
\docType{class}
\name{FLmse-class}
\alias{FLmse-class}
\alias{FLmse}
\alias{FLmse-methods}
\alias{tracking}
\alias{tracking-methods}
\alias{tracking,FLmse-method}
\alias{tracking<-}
\alias{tracking<--methods}
\alias{control}
\alias{control-methods}
\alias{control,FLmse-method}
\alias{control<-}
\alias{control<--methods}
\alias{oem}
\alias{oem-methods}
\alias{oem,FLmse-method}
\alias{oem<-}
\alias{oem<--methods}
\alias{args,FLmse-method}
\alias{args<-,FLmse,list-method}
\title{S4 class \code{FLmse}}
\usage{
FLmse(...)

FLmse(...)

tracking(object, ...)

\S4method{tracking}{FLmse}(object)

tracking(object) <- value

control(object, ...)

\S4method{control}{FLmse}(object)

control(object) <- value

oem(object, ...)

\S4method{oem}{FLmse}(object)

oem(object) <- value

\S4method{args}{FLmse}(name)

\S4method{args}{FLmse,list}(object) <- value
}
\arguments{
\item{...}{additional argument list that might never be used}

\item{object}{object of relevant class (see signature of method)}

\item{value}{the new object}
}
\description{
The \code{FLmse} class stores information relative to the MSE's management procedure'.
}
\section{Slots}{

\describe{
\item{om}{\code{FLom} with the operating model.}
\item{tracking}{\code{FLQuant} with record of decisions made during the mp cycle.}
\item{args}{\code{list} with assorted arguments required to run the MSE cycle.}
}
}

\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity. If an unnamed \code{FLQuant} object is provided, this is used
for sizing, but not for populating any slot.
}

\examples{

}
