% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLoem-class.R, R/FLom-class.R
\docType{class}
\name{FLoem-class}
\alias{FLoem-class}
\alias{FLoem}
\alias{FLoem-methods}
\alias{observations}
\alias{observations-methods}
\alias{observations,FLoem-method}
\alias{observations<-}
\alias{observations<--methods}
\alias{observations<-,FLoem,list-method}
\alias{deviances}
\alias{deviances-methods}
\alias{deviances,FLoem-method}
\alias{deviances<-}
\alias{deviances<--methods}
\alias{show,FLoem-method}
\alias{iter,FLoem-method}
\alias{combine,FLoem,FLoem-method}
\alias{fleetBehaviour<-,FLom,mseCtrl-method}
\alias{projection<-,FLom,mseCtrl-method}
\title{S4 class \code{FLoem}}
\usage{
FLoem(...)

FLoem(...)

observations(object, ...)

\S4method{observations}{FLoem}(object, ...)

observations(object) <- value

\S4method{observations}{FLoem,list}(object) <- value

deviances(object, ...)

\S4method{deviances}{FLoem}(object, ...)

deviances(object) <- value

\S4method{deviances}{FLoem}(object, ...)

\S4method{show}{FLoem}(object)

\S4method{iter}{FLoem}(obj, iter)

\S4method{combine}{FLoem,FLoem}(x, y, ...)

\S4method{fleetBehaviour}{FLom,mseCtrl}(object) <- value

\S4method{projection}{FLom,mseCtrl}(object) <- value
}
\arguments{
\item{...}{additional argument list that might never be used}

\item{object}{object of relevant class (see signature of method)}

\item{value}{the new object}
}
\description{
The \code{FLoem} class stores the information relative to the observation error model of the MSE.
}
\section{Slots}{

\describe{
\item{catch.dev}{\code{FLQuant}  with catch-at-age deviances}
\item{index.qdev}{\code{FLQuants} with catchability deviances for abundance indices}
\item{ctrl}{\code{mseCtrl} with control parameter for the oem}
}
}

\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity. If an unnamed \code{FLQuant} object is provided, this is used
for sizing, but not for populating any slot.
}

\examples{
data(cod)

}
