% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLombf-class.R
\name{FLombf}
\alias{FLombf}
\alias{FLombf-methods}
\alias{append,FLombf,FLombf-method}
\title{Append two FLombf objects along the year dimension}
\usage{
FLombf(...)

\S4method{append}{FLombf,FLombf}(x, values, after = dims(values)$minyear)
}
\arguments{
\item{...}{additional argument list that might never be used}

\item{x}{An \code{FLombf} object, used up to year \code{after}.}

\item{values}{An \code{FLombf} object, used from year \code{after + 1}.}

\item{after}{Integer year. The last year taken from \code{x}. Defaults to
the last year of \code{x}.}

\item{object}{object of relevant class (see signature of method)}
}
\value{
A new \code{FLombf} object spanning the years of \code{x} up to
\code{after} joined with the years of \code{values} from \code{after + 1}.
}
\description{
Joins two \code{FLombf} objects along the year axis: the first object is
windowed to \code{after}, and the second from \code{after + 1} to its last
year. The two windowed pieces are then combined using \code{FLCore::append}
on the underlying \code{FLBiols} and \code{FLFisheries} slots.
}
\examples{
# Split an FLombf at year 2020 and re-join
om1 <- window(om, end=2020)
om2 <- window(om, start=2021)
om3 <- append(om1, om2, after=2020)

}
