% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwd.R
\name{fwd,FLom,missing,fwdControl-method}
\alias{fwd,FLom,missing,fwdControl-method}
\title{res <- fwd(om, control=fwdControl(year=2018:2039, quant="f",
value=rep(c(refpts(om)$FMSY), 22)))}
\usage{
\S4method{fwd}{FLom,missing,fwdControl}(object, control, effort_max = 4, deviances = residuals(sr(object)), ...)
}
\description{
res <- fwd(om, control=fwdControl(year=2018:2039, quant="f",
value=rep(c(refpts(om)$FMSY), 22)))
}
