% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwd.R
\name{fwd,FLom,missing,missing-method}
\alias{fwd,FLom,missing,missing-method}
\title{res <- fwd(om, fbar=FLQuant(c(refpts(om)$FMSY), dimnames=list(year=2018:2039, iter=1:250)))}
\usage{
\S4method{fwd}{FLom,missing,missing}(object, effort_max = 4, deviances = residuals(sr(object)), ...)
}
\description{
res <- fwd(om, fbar=FLQuant(c(refpts(om)$FMSY), dimnames=list(year=2018:2039, iter=1:250)))
}
