% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLombf-class.R
\name{hr,FLombf-method}
\alias{hr,FLombf-method}
\title{hr method for FLombf}
\usage{
\S4method{hr}{FLombf}(object)
}
\arguments{
\item{object}{An \code{FLombf} object.}
}
\value{
An FLQuant (or aggregated FLQuant) that is the sum of the partial
harvest rates across fisheries.
}
\description{
S4 method for \code{hr(object)} when \code{object} is an \code{FLombf}.
The method sums partial harvest-rate contributions across fisheries to
produce a total harvest-rate estimate.
}
