% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{ices.hcr}
\alias{ices.hcr}
\title{Evaluate the chosen HCR function}
\usage{
ices.hcr(stk, fmin, ftrg, blim, bsafe, genArgs, tracking)
}
\arguments{
\item{stk}{The perceived FLStock.}

\item{method}{Name of the chosen HCR function.}

\item{EFF}{Effort array (if effort management is being used).}

\item{EFF0}{Tracking array.}

\item{control}{A list with the elements fmin, ftrg, blim, bsafe and ssb_lag, all of which are numeric.}

\item{ay}{The year for which the target F is set, based on the SSB in year (ay - control$ssb_lag).}
}
\description{
Evaluate the chosen HCR function using the current stock perception and a control
}
