% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLombf-class.R
\name{index.hat,FLIndexBiomass,FLStock-method}
\alias{index.hat,FLIndexBiomass,FLStock-method}
\title{Predicted index of abundance from abundance estimates}
\usage{
\S4method{index.hat}{FLIndexBiomass,FLStock}(object, stock)
}
\arguments{
\item{object}{An FLIndexBiomass object.}

\item{stock}{An FLStock object with estimated abundances, \emph{stock.n}.}
}
\value{
An FLQuant for the precited index of abundance in biomass.
}
\description{
Predicted index of abundance from abundance estimates
}
