% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{indicator.hcr}
\alias{indicator.hcr}
\title{A indicator based HCR}
\usage{
indicator.hcr(stk, hcrpars, genArgs, tracking)
}
\arguments{
\item{stk}{The perceived FLStock.}

\item{genArgs}{A list with generic arguments to be used by the function if needed.}

\item{tracking}{The tracking matrix.}

\item{itrg}{The target for the indicator.}
}
\description{
Get indicator to target. The control argument is a list of parameters used by the HCR.
}
