% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om.R
\name{initiate}
\alias{initiate}
\title{Initializes a population for a given virgin biomass.}
\usage{
initiate(biol, B0, h = 0.75)
}
\arguments{
\item{B0}{Population biology, \code{FLBiol}.}

\item{biom}{Initial or virgin biomass.}
}
\value{
An updated \code{FLBiol} with abundances set in year 1 to match the requested biomas.
}
\description{
The \code{FLBiol} object to be initiated requires slots to be filled up for the
mean weight-at-age (\code{wt}), natural mortality (\code{m}), time of spawning (\code{spwn})
and maturity (\code{mat}).
}
\examples{
data(ple4.biol)
initiate(ple4.biol, B0=450000)
#
initiate(propagate(ple4.biol, 100), B0=runif(100, 3e5, 5e5))
}
