% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{labelPerformance}
\alias{labelPerformance}
\alias{setLabelPerformance}
\title{labelPerformance}
\usage{
labelPerformance(dat, labels = NULL)

setLabelPerformance(file = "model/performance.dat.gz", labels)
}
\arguments{
\item{dat}{A performance statistics table, as returned by performance()}

\item{labels}{Labels to be inserted, as vector or data.frame/data.table}

\item{file}{File where performance table is stored, default: 'model/performance.dat.gz'}
}
\value{
A labelled performance statistics data.table

Invisible updates the table in file
}
\description{
Creates a label column in a performance table

FUNCTION_DESCRIPTION
}
\details{
\itemize{
\item 'numeric'
\item vector
\item data.frame or data.table
\item NULL
}

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
