% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\docType{methods}
\name{mcN}
\alias{mcN}
\title{Compute number of necessary Monte Carlo runs}
\usage{
\S4method{mcN}{FLQuant}(x, s, E = 5, z = 1.96, start = 100, by = 1)
}
\arguments{
\item{E}{The required percentage error of the mean, \emph{numeric}.}

\item{z}{Value of Z for a given confidence level for a normally distributed random variable, default is 1.96 for a 95% CI, \emph{numeric}.}
}
\description{
E. Bukaci et al.Int. Journal of Engineering Research and Applications www.ijera.com ISSN: 2248-9622, Vol. 6, Issue 6, (Part - 3) June 2016, pp.60-64
}
\examples{
\dontrun{
 data(ple4) 
 ssb <- rlnorm(2000, log(ssb(ple4)), 0.5)
 itse <- mcN(ssb) 
 plot(se~iters, itse, type='l')
 }
}
