% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{model_search}
\alias{model_search}
\title{Stepwise model search}
\usage{
model_search(
  x,
  lists = NULL,
  min_model = Freq ~ .,
  max_model = Freq ~ .^2,
  start_model = NULL,
  year = NULL,
  degree_year = NULL
)
}
\arguments{
\item{x}{data frame with the lists and (optionally) covariates in the form of a
contingency table with the variable \code{Freq} containing the observed frequencies,
or in the form of a data frame with individual observations. See details for the
specifications of the variables.}

\item{lists}{numeric vector with the column numbers of \code{x} containing the lists.}

\item{min_model}{formula specifying the minimal model considered in the model search.
Defaults to the main effects models \code{Freq ~ .}.}

\item{max_model}{formula specifying the maximal model considered in the model search.
Defaults to the first-order interaction model \code{Freq ~ .^2}.}

\item{start_model}{formula specifying the first model in the model search. The model
should be inside the range of models specified by \code{min_model} and \code{max_model}.
Defaults to the model specified by \code{min_model}.}

\item{year}{character string with the name of the covariate in \code{x} containing the
year of the observations. The default \code{NULL} applies when the covariate is absent.}

\item{degree_year}{degree of the orthogonal polynomials (1 = linear, 2 = quadratic,
3 = cubic, etc.) to be created for \code{year}. The default \code{NULL} creates dummy
variables.}
}
\value{
Displays the degrees of freedom, deviance, AIC, AICc, BIC and population
size estimates of the models in the search path, and returns a
list with the following objects:
\item{models}{the table with the fit statistics of the fitted models.}
\item{formulas}{the formulas of the fitted models.}
\item{coefs}{the parameter estimates of the fitted models.}
\item{fits}{a list with the fitted frequencies of the models in the search path.}
\item{minima}{a vector with the respective unscaled minima of the AIC, AICc and BIC.}
\item{d}{the data frame \code{x} in the form of a contingency table.}
\item{d_year}{the data frame \code{d} with \code{year} (if present) coded as polynomial.}
\item{obs}{a vector distinguishing between observations and structural zeros in \code{d}.}
\item{lists}{a vector with the column numbers of \code{d} with the lists.}
\item{year}{a scalar indicating the column number of the variable \code{year} in \code{d}.}
}
\description{
\code{model_search} performs a step-wise log-linear model search, and
provides an overview of the model selection criteria AIC, AICc and BIC and the
estimated population size for the models in the search path.
}
\details{
The lists in \code{x} should be coded 0 and 1, with 0 denoting absence and 1 presence on the
list.

If the observations are made in subsequent, non-overlapping periods of time (e.g. years),
the covariate \code{year} indicates the period the observation was made. The maximum value for
\code{degree_year} is one minus the number of distinct periods.
}
\examples{
# Model search with 1st-order interactions term only, and quadratic contrasts for year

search <- model_search(x = simdat, lists = 1:4, year = "Y", degree_year = 2)
}
