% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{movingF.hcr}
\alias{movingF.hcr}
\title{\link{TODO:description}}
\usage{
movingF.hcr(stk, hcrpars, args, tracking)
}
\arguments{
\item{stk}{\link{TODO:description}}

\item{hcrpars}{\link{TODO:description}}

\item{args}{\link{TODO:description}}

\item{tracking}{\link{TODO:description}}
}
\value{
\link{TODO:description}
}
\description{
\link{TODO:description}
}
