% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oem.R
\name{perfect.oem}
\alias{perfect.oem}
\title{A perfect observation of catch and abundances-at-age.}
\usage{
perfect.oem(om, deviances, observations, args, tracking, biomass = FALSE, ...)
}
\arguments{
\item{om}{An operating model, class \emph{FLom} or \emph{FLombf}.}

\item{deviances}{A named list of observation deviances, class \emph{list}.}

\item{observations}{A list of past observations, extended to the end of \emph{om}, class \emph{list}.}

\item{args}{The mp dimensions arguments, as generated by \code{mp}, class \emph{list}.}

\item{tracking}{Object to track module decisions and outputs, class \emph{FLQuant}.}
}
\value{
A named \emph{list} with elements \emph{stk} (\emph{FLStock}), \emph{idx} (\emph{FLIndices}), \emph{deviances}, \emph{observations} and \emph{tracking}.
}
\description{
This observation error model function generates a set of perfect observations
on catches, biology and abundance. Direct observations are made of the stock,
while a single age-structured index of abundance, in numbers, is created with
a fixed catchability of 0.01. \emph{deviances} on either \emph{stk$catch.n} or
\emph{idx$index}, if given, are applied.
}
\details{
This \emph{oem} function generates a full observation time series every time step,
and does not append them to existing objects in \emph{observations}.
}
\examples{
# On FLom
data(ple4om)
obs <- perfect.oem(om, deviances=NULL, observations=NULL,
  args=list(y0=1957, dy=2017), tracking=FLQuant())
}
