% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa.R
\name{perfect.sa}
\alias{perfect.sa}
\title{A perfect 'estimate' of abundances, catches and harvest.}
\usage{
perfect.sa(stk, idx, args, tracking, ...)
}
\arguments{
\item{stk}{The stock observation generated by \emph{oem}. Class \emph{FLStock}.}

\item{idx}{An observation of chnages in abundance, not used. Class \emph{FLIndices}.}

\item{args}{MSE arguments, class \emph{list}.}

\item{tracking}{Structure for tracking modules outputs.}
}
\value{
A \emph{list} with elements \emph{stk} and \emph{tracking}.
}
\description{
The \emph{FLStock} generated by the call to \emph{oem} is simply passed on in this
function. The estimates of abundance, catches and exploitation will thus be
as precise as the OEM observation.
}
\examples{
# Example dataset
data(sol274)

# Sets up an mpCtrl for catch ~ MSY
ctrl <- mpCtrl(est = mseCtrl(method=perfect.sa),
  hcr = mseCtrl(method=fixedC.hcr, args=list(ctrg=11400)))

# Runs mp between 2021 and 2035
run <- mp(om, control=ctrl, args=list(iy=2021, fy=2035))
}
