% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{pid.hcr}
\alias{pid.hcr}
\title{pid
T_{y+1} = T_{y} * 1 - k1 * |lambda| ^ gamma, lambda < 0
1 + k2 * lambda, lambda >= 0}
\usage{
pid.hcr(
  stk,
  ind,
  ref,
  metric = ssb,
  kp = 0,
  ki = 0,
  kd = 0,
  nyears = 5,
  dlow = NA,
  dupp = NA,
  args,
  tracking
)
}
\arguments{
\item{nyears}{Number of years used in regression of log(stock).}
}
\description{
pid
T_{y+1} = T_{y} * 1 - k1 * |lambda| ^ gamma, lambda < 0
1 + k2 * lambda, lambda >= 0
}
\examples{
data(sol274)
track <- FLQuants(FLQuant(dimnames=list(metric='hcr', year=2000:2005,
  iter=seq(dims(om)$iter))))
 args <- list(ay=2003, data_lag=1, management_lag=1, frq=1, it=1)
#
 pid.hcr(stock(om), ind=FLQuant(), tracking=track, args=args,
 nyears=5, metric=ssb, ref=yearMeans(ssb(om)), kp=0.5, ki=0.01, kd=0.7)
# 
control <- mpCtrl(list(
  est = mseCtrl(method=perfect.sa),
  hcr = mseCtrl(method=pid.hcr,
    args=list(metric=ssb, ref=refpts(om)$SBMSY, kp=0.5, ki=0.01, kd=0.7))))
tes <- mp(om, oem=oem, ctrl=control, args=list(iy=2017))
plot(om, PID=tes)
}
