% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plesim}
\alias{plesim}
\title{FLom object}
\format{
An object of class \emph{FLom}.
}
\description{
Aliquam sagittis feugiat felis eget consequat. Praesent eleifend dolor massa,
vitae faucibus justo lacinia a. Cras sed erat et magna pharetra bibendum quis in
mi. Sed sodales mollis arcu, sit amet venenatis lorem fringilla vel. Vivamus vitae
ipsum sem. Donec malesuada purus at libero bibendum accumsan. Donec ipsum sapien,
feugiat blandit arcu in, dapibus dictum felis.
}
\keyword{datasets}
