% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotTOs}
\alias{plotTOs}
\title{plotTOs(perq, "S1", "T1", "2028") + xlim(c(0, 0.30)) + ylim(c(0.5, 1.5))}
\usage{
plotTOs(data, x, y, year = max(data$year), alpha = 1, size = 0.75,
  colkey = "mp")
}
\description{
plotTOs(perq, "S1", "T1", "2028") + xlim(c(0, 0.30)) + ylim(c(0.5, 1.5))
}
