% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oem.R
\name{sampling.oem}
\alias{sampling.oem}
\title{Samples from an operating model to obtain catch, biology and abundance data}
\usage{
sampling.oem(stk, deviances, observations, args, tracking)
}
\arguments{
\item{stk}{An FLStock object as obtained by the call to \emph{stock(om)}.}

\item{deviances}{A named list of deviances, see Details.}

\item{observations}{A named list of observations, see Details.}

\item{args}{Options and arguments passed on by \emph{mp()}.}

\item{tracking}{The tracking object.}
}
\value{
A named list with elements \emph{stk}, \emph{idx}, \emph{observations} and \emph{tracking}.
}
\description{
This observation error model (OEM) function mimics the most common data
collection regime, in which catch-at-age and biology is sampled from the
population, and one or more indices of abundance are derived from surveys
or CPUE data.
}
\details{
The FLStock object passed to \emph{sampling.oem} by the \emph{mp} function is
simplified to match the dimensions of that present in the \emph{observations}
slot.
}
\examples{
data(sol274)
# Generate samples from year 2000:2016
sampling.oem(stock(om), deviances=deviances(oem),
  observations=observations(oem),
  args=list(y0=2000, dy=2021, frq=1), tracking=FLQuant())
}
\seealso{
\link{mp}
}
\author{
Iago Mosqueira (WUR) & Ernesto Jardim (MSC).
}
\keyword{function}
