% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{selectMetric}
\alias{selectMetric}
\title{Select and/or Compute a Metric from the \emph{stk} and \emph{ind} inputs}
\usage{
selectMetric(metric = "missing", stk, ind, ...)
}
\arguments{
\item{metric}{A metric to use, which can be one of the following:
\itemize{
\item \code{missing}: Defaults to the first element of the \code{ind} object if only one element is present.
\item \code{character}: The name of a metric in \code{ind} to extract, or the name of a function to compute the metric.
\item \code{function}: A function to compute the metric with \emph{stk} as input.
}}

\item{stk}{The stock object, used for computing the metric (if applicable).}

\item{ind}{A FLQuants object containing potential metrics, used for extraction based on \code{metric}.}

\item{...}{Additional arguments passed to the function \code{metric} (if it is a function or callable).}
}
\value{
The selected or computed metric, either extracted from \code{ind} or computed using \code{stk} and \code{metric}.
}
\description{
A metric, defined here as a time series, commonly age-aggregated, is computed or extracted
from the input FLStock (\emph{stk}) and FLQuants (\emph{ind}). These have been returned by the call to the
\emph{est}inmation step in a call to \code{mp()}.
}
\details{
If \emph{metric} is a character string and matches a name in \emph{ind}, then that \emph{FLQuant} element
is returned. Otherwise, or if \emph{metric} is a function, it is called on \emph{stk}. See examples below.
}
\examples{
data(ple4)
# Computes 'catch' metric from 'stk', 'ind' is empty
selectMetric("catch", stk=ple4, ind=FLQuants())

# Computes own rfunction (ratio of discards to landings) as metric from 'stk'
selectMetric(function(x) discards(x) / landings(x), stk=ple4, ind=FLQuants())

# Returns 'catch' metric from 'ind' (defined as log), takes precedence over 'stk'
selectMetric("catch", stk=ple4, ind=FLQuants(catch=log(catch(ple4))))

# Any function available for 'stk' works
selectMetric(ssb, stk=ple4, ind=FLQuants())
}
\author{
Iago Mosqueira (WUR)
}
