% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{split.is}
\alias{split.is}
\title{Split a biol target across fleets by setting relative 'quant' proportions}
\usage{
split.is(stk, ctrl, split, quant = unique(ctrl$quant)[1], args, tracking)
}
\description{
Split a biol target across fleets by setting relative 'quant' proportions
}
\examples{
data(ple4)
}
