% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/statistics.R
\docType{data}
\name{statistics}
\alias{statistics}
\title{Example set of performance statistics}
\format{
An object of class list.

A named list with elements of class list, each containing three elements: a formula, a long name ('name'), and a description ('desc'), both of type 'character':
\describe{
\item{SB}{SB: Mean spawner biomass}
\item{SB0}{SB/SB[0]: Mean spawner biomass relative to unfished}
\item{minSB0}{min(SB/SB[0]): Minimum spawner biomass relative to unfished}
\item{SBMSY}{SB/SB[MSY]: Mean spawnwer biomass relative to SBMSY}
\item{F}{F: Mean fishing mortality}
\item{Ftarget}{F/F[target]: Mean fishing mortality relative to target}
\item{FMSY}{F/F[MSY]: Mean fishing mortality relative to FMSY}
\item{green}{P(Green): Probability of being in Kobe green quadrant}
\item{orange}{P(Orange): Probability of being in Kobe orange quadrant}
\item{yellow}{P(Yellow): Probability of being in Kobe yellow quadrant}
\item{red}{P(Red): Probability of being in Kobe red quadrant}
\item{PSBMSY}{P(SB>=SB[MSY]): Probability of SB greater or equal to SBMSY}
\item{PSBlim}{P(SB>SB[limit]): Probability that spawner biomass is above SBlim}
\item{PSB20B0}{P(SB > 0.20 \%*\% SB[0]): Probability that spawner biomass is above 20\% SB\link{0}}
\item{risk1}{mean(P(SB<B[limit])): ICES Risk 1, mean probability that spawner biomass is below Blim}
\item{risk2}{once(P(SB<B[limit])): ICES Risk 2, probability that spawner biomass is above Blim once}
\item{risk3}{max(P(SB>B[limit])): ICES Risk 3, max probability that spawner biomass is above Blim}
\item{C}{mean(C): Mean catch over years}
\item{CMSY}{C/MSY: Mean proportion of MSY}
\item{AAVC}{AAV(C): Average annual variability in catch}
\item{IACC}{IAC(C): Percentage inter-annual change in catch}
\item{PC0}{P(shutdown): Probability of fishery shutdown, defined as catch less than 10\% of MSY}
}
}
\usage{
data(statistics)
}
\description{
A set of performance statistics is provided, coded in the format that
\code{performance()}requires. The statistics included in this list are as
follows:

A list containing a large number of performance statistics:
}
\details{
\itemize{
\item SB0: Mean spawner biomass relative to unfished.
\item SBMSY: Mean spawnwer biomass relative to SBMSY.
\item Ftarget: Mean fishing mortality relative to target.
\item FMSY: Mean fishing mortality relative to FMSY.
\item green: Probability of being in Kobe green quadrant.
\item red: Probability of being in Kobe red quadrant.
\item PSBMSY: Probability of SB greater or equal to SBMSY.
\item PBlim: Probability that spawner biomass is above Blim.
\item risk1: ICES Risk 1, mean probability that spawner biomass is below Blim.
\item risk2: ICES Risk 2, probability that spawner biomass is above Blim once.
\item risk3: ICES Risk 3, max probability that spawner biomass is above Blim.
\item C: Mean catch over years.
\item VarC: Catch variability.
\item PC0: Probability of fishery shutdown.
}

Each indicator is itself a list object, with three elements, the first two
of them compulsory:
\itemize{
\item An unnamed element of class \emph{formula}, e.g. \code{yearMeans(SB/SB0)}.
\item name: A short name to be output on tables and plots, of class character,
e.g. "SB/SB\link{0}".
\item desc: A longer description of the indicator, of class character, e.g. "Mean
spawner biomass relative to unfished"
}
}
\keyword{datasets}
