% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{target.hcr}
\alias{target.hcr}
\title{Target-based harvest control rule to adjust input or output from CPUE}
\usage{
target.hcr(
  ind,
  lim,
  target,
  r = 1,
  metric = "mlc",
  output = "catch",
  nyears = 3,
  args,
  tracking
)
}
\arguments{
\item{PARAM}{Lorem ipsum dolor sit amet}
}
\value{
RETURN Lorem ipsum dolor sit amet
}
\description{
\deqn{}{}
}
\references{
Hoshino, E., Hillary, R., Davies, C., Satria, F., Sadiyah, L., Ernawati, T., and Proctor, C. 2020.  Development of pilot empirical harvest strategies for tropical tuna in indonesian archipelagic waters: case studies of skipjack and yellowfin tuna. \emph{Fisheries Research}, 227:105539, doi:10.1016/j.fishres.2020.105539.
}
\seealso{
\link{FLComp}
}
\author{
The FLR Team
}
\keyword{classes}
