% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{tunebisect}
\alias{tunebisect}
\title{Tune Management Procedures Using Bisection Method}
\usage{
tunebisect(
  om,
  oem = NULL,
  control,
  statistic,
  metrics = NULL,
  args,
  tune,
  prob = 0.5,
  tol = 0.01,
  maxit = 12,
  years = ac(seq(args$iy + 1, args$fy - 1)),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{om}{An object representing the Operating Model, of class \linkS4class{FLom} or \linkS4class{FLombf}.}

\item{oem}{Observation Error Model, class \linkS4class{FLoem} or missing, the default.}

\item{control}{A control object containing the settings and parameters for the Management Procedure, class  \linkS4class{mpCtrl}.}

\item{statistic}{A named list of length 1 defining the statistic used for tuning. The list must include:
\describe{
\item{name}{The name of the statistic.}
\item{formula}{The formula used to compute the statistic.}
\item{desc}{A description of the statistic.}
}}

\item{metrics}{Optional. A set of metrics used for performance evaluation; defaults to \code{NULL} to use the default metrics for the class of the \code{om} object.}

\item{args}{A list of arguments for running the Management Procedure. Must include  \code{iy} (initial year).}

\item{tune}{A named list specifying the parameter of the HCR to tune and its range.  The list must include  the parameter name and the minimum and maximum values as a vector of length 2.}

\item{prob}{The target probability for tuning; defaults to \code{0.5}. Must be a value between 0 and 1.}

\item{tol}{Tolerance for the difference between the computed and target probabilities; defaults to \code{0.01}.}

\item{maxit}{Maximum number of iterations, defaults to 12.}
}
\description{
The \code{tunebisect} function is designed to tune a single parameter of a single
modules in a Management Procedure (MP), typically in the Harvest Control Rule (hcr).
It uses a bisection algorithm to iteratively adjust the parameter value unitl it
achieves specified probability value for a given management objective overn a
selected time frame. For example, it can try to find a value for the \code{target}
argument of a \code{hocketstick.hcr} that obtains a 60\% mean probability of SSB being above
SSB\link{MSY} over the 2032 to 2042 period.
}
\details{
Two initial MP runs are carried out employing the argument values provided in \code{tune}.
They need to return probabilities that are on both sides of the expected one. If not,
the function will return both runs for inspection. The function should then be rerun
with an wider range of values, or the time-frame extended.
}
\examples{
# dataset contains both OM (FLom) and OEM (FLoem)
data(plesim)
# choose sa and hcr
control <- mpCtrl(list(
  est = mseCtrl(method=perfect.sa),
  hcr = mseCtrl(method=hockeystick.hcr, args=list(lim=0,
  trigger=14000, target=0.18))))
# load statistics
data(statistics)
tun <- tunebisect(om, oem=oem, control=control, args=list(iy=2021, fy=2035),
tune=list(target=c(0.12, 0.32)),
metrics=list(SB=ssb), statistic=statistics['PSBMSY'], years=2025:2034)
# Plot tuned MP
plot(om, tun)
}
\references{
{
Burden, Richard L.; Faires, J. Douglas (2016), "2.1 The Bisection Algorithm", Numerical Analysis (10th ed.), Cenage Learning, ISBN 978-1-305-25366-7
}
}
