% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{westeros}
\alias{westeros}
\title{westeros.}
\format{A data frame with 1043 rows and 9 variables:
\describe{
  \item{R1}{population register coded 1 if the case is observed, and 0 otherwise}
  \item{R2}{population register coded 1 if the case is observed, and 0 otherwise}
  \item{R3}{population register coded 1 if the case is observed, and 0 otherwise}
  \item{R4}{population register coded 1 if the case is observed, and 0 otherwise}
  \item{X1}{a covariate with 2 levels}
  \item{X2}{a covariate with 2 levels}
  \item{X3}{a covariate with 2 levels}
  \item{X4}{a covariate with 3 levels}
  \item{Y}{a variable denoting the year the observation was made}
}}
\usage{
westeros
}
\description{
A simulated dataset containing MSE data.
}
\keyword{datasets}
