% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{writePerformance}
\alias{writePerformance}
\title{Write performance table to file}
\usage{
writePerformance(dat, file = "model/performance.dat.gz", overwrite = FALSE)
}
\arguments{
\item{dat}{data.table with performance statistics.}

\item{file}{file to write to, defaults to 'model/performance.dat.gz' a compressed text file.}
}
\value{
file name, invisibly.
}
\description{
Write performance table to file
}
\author{
Iago Mosqueira, WMR
}
\keyword{utilities}
