% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{boot_mse}
\alias{boot_mse}
\title{Estimation of 95\% bootstrap confidence intervals.}
\usage{
boot_mse(object, modelnr = NULL, rep = 1000, seed = 1)
}
\arguments{
\item{object}{an object created with the function \code{\link{model_search}}.}

\item{modelnr}{the number of the model in \code{object} that is to be bootstrapped.
If left unspecified, the  model with the lowest BIC will be selected.}

\item{rep}{the number of bootstrap replications.}

\item{seed}{a scalar to set the random seed.}
}
\value{
A lists with tables and, if the variable \code{year} is specified, with plots.
\item{tables}{table(s) with the number of observations, the population size estimate with
the 95\% bootstrap confidence interval. If covariates are present, the tables are presented
for the levels of the covariate and, with more than one covariae, for the combinations of
the levels of pairs of covariates. If the variable \code{year} is specified, these
tables are presented by year.}
\item{plots}{plots of the population size estimates and their 95\% confidence intervals,
split out by the levels of the covariates and with the variable \code{year} on the x-axis.
Only available if the variable \code{year} is specified.}
}
\description{
\code{boot_mse} performs a parametric bootstrap for a selected model.
}
