% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{catchSSB.hcr}
\alias{catchSSB.hcr}
\title{A HCR to set total catch based on SSB depletion level}
\usage{
catchSSB.hcr(
  stk,
  dtarget = 0.4,
  dlimit = 0.1,
  lambda = 1,
  MSY,
  dtaclow = 0.85,
  dtacupp = 1.15,
  yrs = 1,
  metric = "ssb",
  args,
  tracking
)
}
\arguments{
\item{stk}{The perceived FLStock.}

\item{MSY}{Assumed or estimated MSY.}

\item{args}{MSE arguments, class \emph{list}.}

\item{tracking}{Structure for tracking modules outputs.}

\item{dtarget=0.40}{Depletion level from which catch is decreased.}

\item{dlimit=0.10}{Depletion level at which fishing is stopped.}

\item{lambda=1}{Multiplier for MSY level.}

\item{dtaclow=0.85}{Maximum proportional decrease in allowable catch.}

\item{dtacupp=1.15}{Maximum proportional increase in allowable catch.}
}
\value{
A \emph{list} with elements \emph{ctrl}, of class \emph{fwdControl}, and \emph{tracking}.
}
\description{
A HCR to set total catch based on SSB depletion level
}
\examples{
data(sol274)
catchSSB.hcr(stock(om), MSY=140000, tracking=FLQuant(),
args=list(ay=2018, data_lag=1, management_lag=1, frq=1))
# APPLY hcr over a range of dtarget values
lapply(seq(0.30, 0.80, by=0.1), function(x) {
  catchSSB.hcr(stock(om), MSY=140000, dtarget=x,
  args=list(ay=2018, data_lag=1, management_lag=1, frq=1),
  tracking=FLQuant())$ctrl } )
}
