% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{catchSSB.hcr}
\alias{catchSSB.hcr}
\title{A HCR to set catch based on SSB}
\usage{
catchSSB.hcr(
  stk,
  dtarget = 0.4,
  dlimit = 0.1,
  lambda = 1,
  MSY,
  ssb_lag = 1,
  dtaclow = 0.85,
  dtacupp = 1.15,
  ay,
  tracking
)
}
\description{
A HCR to set catch based on SSB
}
\examples{
data(ple4)
# APPLY hcr
catchSSB.hcr(ple4[,ac(1957:1996)], dtarget=0.40, dlimit=0.10, lambda=1,
  MSY=95000, ay=1995, tracking=FLQuant())
# APPLY hcr over a range of dtarget
lapply(seq(0.30, 0.90, by=0.1), function(x) {
   catchSSB.hcr(ple4[,ac(1957:1996)], dtarget=x, dlimit=0.10, lambda=1,
  MSY=95000, ay=1995, tracking=FLQuant())$ctrl})
}
