% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{computeFp05}
\alias{computeFp05}
\title{Calculates the Fbar value giving a probability of ssb being Blim of 5\%}
\usage{
computeFp05(
  stock,
  sr,
  SBlim,
  range = c(0.01, 0.75),
  nyears = 3,
  sigmaR = 0.5,
  rho = 0.43,
  its = 500,
  verbose = TRUE
)
}
\description{
Calculates the Fbar value giving a probability of ssb being Blim of 5\%
}
\examples{
data(ple4)
sr <- predictModel(model=bevholt, params=FLPar(a=1.4e6, b=1.5e5))
fp05 <- computeFp05(ple4, sr, SBlim=150000, its=300, range=c(0.10,0.40))
# RUN projection for Fp.05 value
proj <- fwd(propagate(stf(ple4, nyears=100), 300), sr=sr,
  fbar=FLQuant(fp05, dimnames=list(year=2018:2117)),
  deviances=ar1rlnorm(rho=0.43, years=2018:2117, iters=300, meanlog=0,
  sdlog=0.5))
plot(ssb(proj), prob=c(0.01, 0.25, 0.50, 0.75, 0.99)) +
  geom_hline(yintercept=150000)
}
