% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{cpue.hcr}
\alias{cpue.hcr}
\title{cpue.hcr}
\usage{
cpue.hcr(
  stk,
  ind,
  k1,
  k2,
  k3,
  k4,
  target = 1,
  dtaclow = 0.85,
  dtacupp = 1.15,
  args,
  tracking
)
}
\description{
cpue.hcr
}
\examples{
data(ple4om)
cpue.hcr(FLQuants(stock(om)), k1=0.1, k2=0.2, k3=0.1, k4=0.1, args=list(ay=1990),
 tracking=FLQuants(FLQuant(c(0.5, 0.8), dimnames=list(metric=c("cpue.slope",
 "cpue.mean"), year=1990))))
}
