% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{cpue.hcr}
\alias{cpue.hcr}
\title{cpue.hcr}
\usage{
cpue.hcr(
  stk,
  ind,
  k1 = 0.2,
  k2 = 0.2,
  k3 = 0.2,
  k4 = 0.2,
  target = 1,
  dtaclow = 0.85,
  dtacupp = 1.15,
  initac = NULL,
  slope = "slope",
  mean = "mean",
  args,
  tracking
)
}
\description{
cpue.hcr
}
\examples{
data(sol274)
ind <- cpue.ind(stock(om), FLIndices(CPUE=FLIndexBiomass(index=ssb(om))),
  args=list(ay=2000, data_lag=1),
  tracking=FLQuant(dimnames=list(metric="ind", year=2000, iter=1:100)))
cpue.hcr(stk=stock(om), ind=ind$ind, k1=0.1, k2=0.2, k3=0.1, k4=0.1,
  args=list(ay=2000, frq=1, management_lag=1),
  tracking=FLQuants(sol174=FLQuant(1000, dimnames=list(metric="hcr",
  year=2000))))
}
