% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{fixedC.hcr}
\alias{fixedC.hcr}
\title{A fixed target C}
\usage{
fixedC.hcr(stk, ctrg, args, tracking)
}
\arguments{
\item{stk}{The perceived FLStock.}

\item{control}{A list with the element ftrg (numeric).}
}
\description{
No matter what get C = Ctrg
The control argument is a list of parameters used by the HCR.
}
\examples{
# Example dataset
data(sol274)

# Sets up an mpCtrl for catch ~ MSY
ctrl <- mpCtrl(est = mseCtrl(method=perfect.sa),
  hcr = mseCtrl(method=fixedC.hcr, args=list(ctrg=11400)))

# Runs mp between 2021 and 2035
run <- mp(om, control=ctrl, args=list(iy=2021, fy=2035))
}
