% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{fixedC.hcr}
\alias{fixedC.hcr}
\title{A fixed target C}
\usage{
fixedC.hcr(stk, ctrg, args, tracking)
}
\arguments{
\item{stk}{The perceived FLStock.}

\item{control}{A list with the element ftrg (numeric).}
}
\description{
No matter what get C = Ctrg
The control argument is a list of parameters used by the HCR.
}
\examples{
data(sol274)
fixedC.hcr(stock(om), ctrg=50000, args=list(ay=2017, management_lag=1,
  frq=1, it=100), tracking=FLQuant())
}
