% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{fixedF.hcr}
\alias{fixedF.hcr}
\title{A fixed target f}
\usage{
fixedF.hcr(stk, ftrg, args, tracking)
}
\arguments{
\item{stk}{The perceived FLStock.}

\item{control}{A list with the element ftrg (numeric).}
}
\description{
No matter what get F = Ftarget
The control argument is a list of parameters used by the HCR.
}
\examples{
data(sol274om)
fixedF.hcr(stock(om), ftrg=0.15, args=list(ay=2017, management_lag=1,
  frq=1), tracking=FLQuant())
}
