% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{fixedF.hcr}
\alias{fixedF.hcr}
\title{A fixed target f}
\usage{
fixedF.hcr(stk, ftrg, ay, tracking)
}
\arguments{
\item{stk}{The perceived FLStock.}

\item{control}{A list with the element ftrg (numeric).}
}
\description{
No matter what get F = Ftarget
The control argument is a list of parameters used by the HCR.
}
