% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLoem-class.R
\docType{class}
\name{FLoem-class}
\alias{FLoem-class}
\alias{FLoem}
\alias{FLoem-methods}
\alias{observations}
\alias{observations-methods}
\alias{observations,FLoem-method}
\alias{observations<-}
\alias{observations<--methods}
\alias{observations<-,FLoem,missing,list-method}
\alias{observations<-,FLoem,ANY,FLStock-method}
\alias{deviances,FLoem-method}
\alias{deviances<-,FLoem,list-method}
\alias{show,FLoem-method}
\alias{iter,FLoem-method}
\alias{combine,FLoem,FLoem-method}
\title{Specification for the observation error model (OEM).}
\usage{
FLoem(...)

FLoem(...)

observations(object, ...)

\S4method{observations}{FLoem}(object, ...)

observations(object, i) <- value

\S4method{observations}{FLoem,missing,list}(object) <- value

\S4method{observations}{FLoem,ANY,FLStock}(object, i) <- value

\S4method{deviances}{FLoem}(object, ...)

\S4method{deviances}{FLoem,list}(object, ...) <- value

\S4method{show}{FLoem}(object)

\S4method{iter}{FLoem}(obj, iter)

\S4method{combine}{FLoem,FLoem}(x, y, ..., check = FALSE)
}
\arguments{
\item{...}{additional argument list that might never be used}

\item{object}{object of relevant class (see signature of method)}

\item{value}{the new object}
}
\description{
The \code{FLoem} class stores the method, arguments and
observations that define the way observations are collected from an operating
model at each time step in the management procedure. This class extends
\emph{mseCtrl} through the addition of two lists used to gather past and new
observations, and deviances to use on each step in the observation process.
}
\section{Slots}{

\describe{
\item{\code{method}}{The function to be run in the module call, class \emph{function}.}

\item{\code{args}}{Arguments to be passed to the method, of class \emph{list}.}

\item{\code{observations}}{Past observations, class \emph{list}.}

\item{\code{deviances}}{Observation deviances, class \emph{list}.}
}}

\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity. If an unnamed \code{FLQuant} object is provided, this is used
for sizing, but not for populating any slot.
}

\examples{
data(sol274)
deviances(oem)
deviances(oem, "stk")
deviances(oem, "stk", "catch.n")
data(sol274)
x <- iter(oem, 1:50)
dims(observations(x)$stk)$iter
dims(deviances(x)$stk$catch.n)$iter
dims(deviances(x)$idx[[1]])$iter
data(sol274)
iter(oem, 1:50)
}
