% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLom-class.R
\name{fwd.om}
\alias{fwd.om}
\title{A method to project the operating model (OM)}
\usage{
fwd.om(om, ctrl, ...)
}
\arguments{
\item{ctrl}{the fwdControl object with objectives and constraints}

\item{object}{the OM as a FLStock}

\item{sr}{a FLSR with the stock-recruitment model}

\item{sr.residuals}{a FLQuant with S/R residuals}

\item{sr.residuals.mult}{logical about residuals being multiplicative}
}
\description{
Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque eleifend
odio ac rutrum luctus. Aenean placerat porttitor commodo. Pellentesque eget porta
libero. Pellentesque molestie mi sed orci feugiat, non mollis enim tristique.
Suspendisse eu sapien vitae arcu lobortis ultrices vitae ac velit. Curabitur id
}
