% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{hockeystick.hcr}
\alias{hockeystick.hcr}
\title{Hockey-stick Harvest Control Rule}
\usage{
hockeystick.hcr(
  stk,
  ind,
  lim,
  trigger,
  target,
  min = 0,
  drop = 0,
  metric = "ssb",
  output = "fbar",
  dlow = NULL,
  dupp = NULL,
  all = TRUE,
  args,
  tracking
)
}
\arguments{
\item{stk}{The OEM observation or SA estimation, an FLStock object.}

\item{ind}{Possible indicators returned by the 'est' step, FLQuants.}

\item{lim}{The lower threshold of the stock metric, below which the control rule
applies the minimum allowable value (\code{min}), numeric or FLQuant.}

\item{trigger}{The stock status or metric value threshold above which the control variable reaches the target value, numeric or FLQuant.}

\item{target}{The output level applied when the stock metric is at or above the \code{trigger} value, numeric or FLQuant.}

\item{min}{Numeric. The minimum allowable control value (e.g., minimum F or catch).}

\item{drop}{Numeric. A stock metric threshold below which the control variable is forced to \code{min} to prevent over-exploitation of severely depleted stocks, numeric  or FLQuant}

\item{metric}{The stock metric to use for the HCR (e.g., \code{"ssb"} for spawning-stock biomass), character or function.}

\item{output}{Character. The output control variable (e.g., \code{"fbar"} for fishing mortality or \code{"catch"} for quotas), character.}

\item{dlow}{A limit for the decrease in the output variable, e.g. 0.85 for a maximum decrease of 15\%, numeric. No limit is applied if NULL.}

\item{dupp}{A limit for the increase in the output variable, e.g. 1.15 for a maximum increase of 15\%, numeric. No limit is applied if NULL.}

\item{all}{If \code{TRUE}, upper and lower limits (\code{dupp} and \code{dlow}) are applied unconditionally, otherwise only when metric > trigger, logical.}

\item{args}{A list containing dimensionality arguments, passed on by mp().}

\item{tracking}{An FLQuant used for tracking indicators, intermediate values, and decisions during MP evaluation.}
}
\value{
A list containing elements 'ctrl', a fwdControl object, and 'tracking'.
}
\description{
A hockey-stick harvest control rule that sets the value of an output between a
minimum and a target, according to that of a metric, compared with a limit and
a trigger.
}
\details{
This function implements a hockey-stick shaped harvest control rule (HCR). It is
commonly used to set F or effort, and sometimes catch, from a measure of stock
status, such as estimated SSB or depletion. The function can take any input,
'metric', if it can computed from an FLStock or is returned by the function run
in the 'est' step. The 'output' argument can be set to any quantity that the OM
projection method can understand, e.g. fbar, effort or catch for 'fwd.om'.
The HCR increases the control variable linearly between a lower threshold (\code{lim})
and a trigger point (\code{trigger}), and sets it to a target level (\code{target}) when the
metric is above the trigger. The decreasing line can be cut at any point ('drop'),
where output values fall to the 'min' level set. See examples below.
The function can also apply optional upper (\code{dupp}) and lower (\code{dlow}) constraints
to changes in the control variable. These constraints can be applied either
unconditionally (\code{all = TRUE}) or only on the stock being above the trigger.
}
\examples{
data(ple4)

# Define arguments, done automatically when called by mp()
args <- list(ay = 2015, iy = 2015, dy=2014, management_lag = 1, mys=2016)
track <- FLQuant(dimnames=list(metric="fbar.hcr", year=2014:2016))

# Apply the control rule
ctrl <- hockeystick.hcr(stk = ple4, ind = FLQuants(), lim = 7e5, trigger = 1e6, 
  target = 0.24, metric = "ssb", output = "fbar", args = args, tracking = track)

# Inspect the control object
print(ctrl)

# Add change limits
ctrl2 <- hockeystick.hcr(stk = ple4, ind = FLQuants(), lim = 7e5, trigger = 1e6, 
  target = 0.24, dlow=0.85, dupp=1.15, metric = "ssb", output = "fbar",
  args = args, tracking = track)

# Inspect the control object
print(ctrl2)
}
