% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{ices.hcr}
\alias{ices.hcr}
\title{The typical HCR used by ICES}
\usage{
ices.hcr(stk, ftrg, sblim, sbsafe, fmin = 0, args, tracking)
}
\arguments{
\item{stk}{The perceived FLStock.}

\item{ftrg}{\link{TODO:description}}

\item{sblim}{\link{TODO:description}}

\item{sbsafe}{\link{TODO:description}}

\item{fmin}{Minimum fishing mortality.}

\item{args}{MSE arguments, class \emph{list}.}

\item{tracking}{Structure for tracking modules outputs.}
}
\value{
A \emph{list} with elements \emph{ctrl}, of class \emph{fwdControl}, and \emph{tracking}.
}
\description{
The typical HCR used by ICES which sets a target F based on the SSB based on 4 parameters: sblim, sbsafe, fmin and ftrg.
F increases linearly between SSB = blim and SSB = bsafe, from F = fmin to F = ftrg.
If:
\itemize{
\item B < Blim, F = Fbycatch;
\item B > trigger, F = Fmsy;
\item B > Blim & B < trigger, F linear between Fbycatch and Fmsy;
\item F = ftrg is the maximum F, F = fmin is the minimum F.
F is set in year ay, based on SSB in year ay - data_lag
}
}
\examples{
data(ple4om)
# Test for year when SSB > bsafe
ices.hcr(stock(om), fmin=0.05, ftrg=0.15, sblim=200000, sbsafe=300000,
  args=list(ay=2018, data_lag=1, management_lag=1), tracking=FLQuant())
# Test for year when SSB < bsafe
ices.hcr(stock(om), fmin=0.05, ftrg=0.15, sblim=200000, sbsafe=300000,
  args=list(ay=1995, data_lag=1, management_lag=1), tracking=FLQuant())
}
