% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{indicator.is}
\alias{indicator.is}
\title{indicator implementation function}
\usage{
indicator.is(stk, ctrl, args, tracking, system = c("output", "input"), ...)
}
\arguments{
\item{stk}{The perceived FLStock.}

\item{ctrl}{control file with HCR decision}
}
\description{
indicator implementation function
}
