% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om.R
\name{initiate}
\alias{initiate}
\title{Initializes a population for a given virgin biomass.}
\usage{
initiate(biol, B0, h = 0.75)
}
\arguments{
\item{biol}{An \code{FLBiol} object to nbe initiated.}

\item{B0}{Initial or virgin biomass.}
}
\value{
An updated \code{FLBiol} with abundances set in the first year to match
the requested biomas.
}
\description{
Abundances at age for a population at virgin conditions at age. An \code{FLBiol}
object is initiated by providing a target total biomass (\code{B0}) and a value
for the stock-recruit steepness (\code{s}). The object requires slots to be
already filled up for the mean weight-at-age (\code{wt}), natural mortality
(\code{m}), time of spawning (\code{spwn}) and maturity at age (\code{mat}).
}
\examples{
data(ple4.biol)
\dontrun{
initiate(ple4.biol, B0=450000)
# Sets up parallel
# plan(multisession, workers=4)
initiate(propagate(ple4.biol, 100), B0=runif(100, 3e5, 5e5))
}
}
