% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\docType{methods}
\name{mcN}
\alias{mcN}
\title{Compute number of necessary Monte Carlo runs}
\usage{
\S4method{mcN}{FLQuant}(x, s, E = 5, z = 1.96, start = 100, by = 1)
}
\arguments{
\item{E}{The required percentage error of the mean, \emph{numeric}.}

\item{z}{Value of Z for a given confidence level for a normally distributed random variable, default is 1.96 for a 95\% CI, \emph{numeric}.}
}
\description{
The \code{mcN} function implements the simple method of Bukaci et al. (2016) to
calculate the number of Monte Carlo (MC) simulations required to obtain
results with a given precision level.
}
\examples{
\dontrun{
 data(ple4) 
 ssb <- rlnorm(2000, log(ssb(ple4)), 0.5)
 itse <- mcN(ssb) 
 plot(se~iters, itse, type='l')
 }
}
\references{
Bukaci, E. Korini, Th., Periku, E., Allkja, S., Sheperi, P. 2016. Number of iterations needed in Monte Carlo Simulation using reliability analysis for tunnel supports.  Int. J. of Eng. Res. and Apps., 6 (6-3): 60-64. <www.ijera.com/papers/Vol6_issue6/Part%20-%203/J0606036064.pdf>
}
