% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{model_search}
\alias{model_search}
\title{Stepwise model search}
\usage{
model_search(x, lists = NULL, k_max = 1, year = NULL,
  degree_year = NULL)
}
\arguments{
\item{x}{a data frame with the incomplete lists and, if present, covariates.
For the specification of the variables in \code{x}, see Details.}

\item{lists}{a vector with the column numbers of \code{x} with the lists.}

\item{k_max}{a scalar restricting the \emph{k}-order interaction terms to \code{k-max}.}

\item{year}{character string with the name of the variable in \code{x} representing the
years of the observations, if present.}

\item{degree_year}{a scalar denoting the polynomial degree of the variable \code{year},
if present.}
}
\value{
Prints a table with the degrees of freedom, deviance, AIC, AICc, BIC and population
size estimates of the fitted models. The AIC, AICc and BIC are scaled in such a way that
there minimum values equal 0. As input argument for the function \code{\link{boot_mse}} a
list with the following objects is returned:
\item{models}{the table with the fit statistics of the fitted models.}
\item{formulas}{the formulas of the fitted models.}
\item{coefs}{the parameter estimates of the fitted models.}
\item{fits}{a list with the fitted frequencies of the models in the search path.}
\item{minima}{a vector with the respective unscaled minima of the AIC, AICc and BIC.}
\item{d}{the data frame \code{x} in the form of a contingency table.}
\item{d_year}{the data frame \code{d} with \code{year} (if present) coded as polynomial.}
\item{obs}{a vector distinguishing between observations and structural zeros in \code{d}.}
\item{lists}{a vector with the column numbers of \code{d} with the lists.}
\item{year}{a scalar indicating the column number of the variable \code{year} in \code{d}.}
}
\description{
\code{model_search} performs a stepwise log-linear model search.
}
\details{
\code{x} should include at least two lists, and the lists should be coded 1 if the
observation is in that list, and 0 if it is not.

\code{x} can be either in an \eqn{n x p} format, where \eqn{n} is the number of observed
individuals and \eqn{p} the number of variables (lists plus covariates), or in an
\eqn{m x p + 1} format, where \eqn{m} is the number of all possible combination of the variable
levels, and an additional variable called \code{Freq} with the frequencies (including the
structural zeros).
}
