% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{model_search}
\alias{model_search}
\title{Stepwise model search}
\usage{
model_search(x, lists = NULL, max_model = Freq ~ .^2, year = NULL,
  degree_year = NULL)
}
\arguments{
\item{x}{a data frame with the incomplete lists and, if present, covariates.
For the specification of the variables in \code{x}, see Details.}

\item{lists}{a vector with the column numbers of \code{x} corresponding to the lists.}

\item{max_model}{a formula specifying the maximal model in the model search. The default
\code{Freq ~ .^2} restricts the maximal model to first-order interactions terms.}

\item{year}{character string with the name of the covariate with the
years of the observation. The default \code{NULL} assumes the absence of such a covariate.}

\item{degree_year}{if \code{year} is specified, a scalar indicating that polynomial
contrasts should be used for \code{year}. The value 1 creates a linear contrast,
with the first year as reference category
the value 2 quadratic contrasts, etc.. The default \code{NULL} creates dummy variables.}
}
\value{
Prints a table with the degrees of freedom, deviance, AIC, AICc, BIC and population
size estimates of the fitted models. The AIC, AICc and BIC are scaled in such a way that
there minimum values equal 0. As input argument for the function \code{\link{boot_mse}} a
list with the following objects is returned:
\item{models}{the table with the fit statistics of the fitted models.}
\item{formulas}{the formulas of the fitted models.}
\item{coefs}{the parameter estimates of the fitted models.}
\item{fits}{a list with the fitted frequencies of the models in the search path.}
\item{minima}{a vector with the respective unscaled minima of the AIC, AICc and BIC.}
\item{d}{the data frame \code{x} in the form of a contingency table.}
\item{d_year}{the data frame \code{d} with \code{year} (if present) coded as polynomial.}
\item{obs}{a vector distinguishing between observations and structural zeros in \code{d}.}
\item{lists}{a vector with the column numbers of \code{d} with the lists.}
\item{year}{a scalar indicating the column number of the variable \code{year} in \code{d}.}
}
\description{
\code{model_search} performs a stepwise log-linear model search.
}
\details{
The data frame \code{x} can be either in \eqn{n x p} format, where \eqn{n} is the number
of observed individuals and \eqn{p} the number of lists and covariates, or in the form of
a contingency the combinations of the variable levels in \code{x} in the rows, and the lists,
covariates and a variable called \code{Freq} with the corresponding frequencies in the columns.

The \code{lists} should be coded 1 if the corresponding individual/combination of variable levels
is/are observed in the list, and 0 if not.

The covariate represented by \code{year} should denote the year in which the observation was made.
}
\examples{
# Model search with 1st-order interactions term only, and quadratic contrasts for year

search <- model_search(x = westeros, lists = 1:4, year = "Y", degree_year = 2)
}
