% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{mp}
\alias{mp}
\title{mp executes a run of a MP}
\usage{
mp(
  om,
  oem = NULL,
  iem = NULL,
  ctrl,
  args,
  scenario = "test",
  tracking = "missing",
  verbose = TRUE,
  parallel = TRUE
)
}
\arguments{
\item{om}{The operating model (OM), an object of class \emph{FLom}.}

\item{oem}{The observation error model (OEM), an object orf class \emph{FLoem}.}

\item{iem}{The implementation error model (IEM), an object of class \emph{FLiem}.}

\item{ctrl}{A control structure for the MP run, an object of class \emph{mpCtrl}.}

\item{args}{MSE arguments, \emph{list}.}

\item{scenario}{Name of the scenario tested in this ru, \emph{character}.}

\item{tracking}{Extra elements to add to the standard tracking \emph{FLQuant}.}

\item{verbose}{Should output be verbose or not, \emph{logical}.}
}
\value{
An object of class \emph{FLmse}.
}
\description{
mp executes a run of a MP
}
\examples{
# [TODO:example]
}
