% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{mp}
\alias{mp}
\title{mp executes a single run of a Management Procedure}
\usage{
mp(
  om,
  oem = NULL,
  iem = NULL,
  ctrl,
  args,
  scenario = "NA",
  tracking = "missing",
  verbose = TRUE,
  parallel = TRUE
)
}
\arguments{
\item{om}{The operating model (OM), an object of class \emph{FLom} or \emph{FLombf}.}

\item{oem}{The observation error model (OEM), an object of class \emph{FLoem}.}

\item{iem}{The implementation error model (IEM), an object of class \emph{FLiem}.}

\item{ctrl}{A control structure for the MP run, an object of class \emph{mpCtrl}.}

\item{args}{MSE arguments, \emph{list}. Only 'iy', the intermediate (starting) year, is required.}

\item{scenario}{Name of the scenario tested in this run, \emph{character}.}

\item{tracking}{Extra elements (rows) to add to the standard tracking \emph{FLQuant} in its first dimensions, \emph{character}.}

\item{verbose}{Should output be verbose or not, \emph{logical}.}
}
\value{
An object of class \emph{FLmse}.
}
\description{
An individual management procedure (MP) is run for a number of years,
on an operating model, observation error model, control and arguments.
}
\examples{
# dataset contains both OM (FLom) and OEM (FLoem)
data(ple4om)
# choose sa and hcr
control <- mpCtrl(list(
  est = mseCtrl(method=perfect.sa),
  hcr = mseCtrl(method=catchSSB.hcr,
    args=list(MSY=140000))))
tes <- mp(om, oem=oem, ctrl=control, args=list(iy=2017))
# 'perfect.oem' is used if none is given
tes <- mp(om, ctrl=control, args=list(iy=2017))
plot(om, TEST=tes)
}
