% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpCtrl-class.R
\docType{class}
\name{mpCtrl-class}
\alias{mpCtrl-class}
\alias{mpCtrl}
\alias{initialize,mpCtrl-method}
\alias{mpCtrl-methods}
\alias{ctrl.est}
\alias{ctrl.est-methods}
\alias{ctrl.est,mpCtrl-method}
\alias{ctrl.est<-}
\alias{ctrl.est<--methods}
\alias{ctrl.est<-,mpCtrl,function-method}
\alias{ctrl.phcr}
\alias{ctrl.phcr-methods}
\alias{ctrl.phcr,mpCtrl-method}
\alias{ctrl.phcr<-}
\alias{ctrl.phcr<--methods}
\alias{ctrl.phcr<-,mpCtrl,function-method}
\alias{ctrl.hcr}
\alias{ctrl.hcr-methods}
\alias{ctrl.hcr,mpCtrl-method}
\alias{ctrl.hcr<-}
\alias{ctrl.hcr<--methods}
\alias{ctrl.hcr<-,mpCtrl,function-method}
\alias{ctrl.is}
\alias{ctrl.is-methods}
\alias{ctrl.is,mpCtrl-method}
\alias{ctrl.is<-}
\alias{ctrl.is<--methods}
\alias{ctrl.is<-,mpCtrl,function-method}
\alias{ctrl.tm}
\alias{ctrl.tm-methods}
\alias{ctrl.tm,mpCtrl-method}
\alias{ctrl.tm<-}
\alias{ctrl.tm<--methods}
\alias{ctrl.tm<-,mpCtrl,function-method}
\alias{show,mpCtrl-method}
\alias{iters,mpCtrl-method}
\alias{exists}
\alias{exists,mpCtrl-method}
\title{S4 class \code{mpCtrl}}
\usage{
\S4method{initialize}{mpCtrl}(.Object, ...)

ctrl.est(object, ...)

\S4method{ctrl.est}{mpCtrl}(object)

ctrl.est(object) <- value

\S4method{ctrl.est}{mpCtrl,`function`}(object) <- value

ctrl.phcr(object, ...)

\S4method{ctrl.phcr}{mpCtrl}(object)

ctrl.phcr(object) <- value

\S4method{ctrl.phcr}{mpCtrl,`function`}(object) <- value

ctrl.hcr(object, ...)

\S4method{ctrl.hcr}{mpCtrl}(object)

ctrl.hcr(object) <- value

\S4method{ctrl.hcr}{mpCtrl,`function`}(object) <- value

ctrl.is(object, ...)

\S4method{ctrl.is}{mpCtrl}(object)

ctrl.is(object) <- value

\S4method{ctrl.is}{mpCtrl,`function`}(object) <- value

ctrl.tm(object, ...)

\S4method{ctrl.tm}{mpCtrl}(object)

ctrl.tm(object) <- value

\S4method{ctrl.tm}{mpCtrl,`function`}(object) <- value

\S4method{show}{mpCtrl}(object)

\S4method{iters}{mpCtrl}(object, iter)

exists(x, where = -1, envir = if (missing(frame)) as.environment(where)
  else sys.frame(frame), frame, mode = "any", inherits = TRUE)
}
\arguments{
\item{...}{additional argument list that might never be used}

\item{object}{object of relevant class (see signature of method)}
}
\description{
The \code{mpCtrl} class defines which moules will be runned and carries the specifications of those modules in \code{mseCtrl} elements.
}
\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity. If an unnamed \code{FLQuant} object is provided, this is used
for sizing, but not for populating any slot.
}

