% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{mps}
\alias{mps}
\title{mps}
\usage{
mps(om, oem = NULL, iem = NULL, ctrl, args, names = NULL, parallel = TRUE, ...)
}
\description{
mps
}
\examples{
data(sol274)
# Set control: sa and hcr
control <- mpCtrl(list(
  est = mseCtrl(method=perfect.sa),
  hcr = mseCtrl(method=hockeystick.hcr, args=list(lim=0,
  trigger=41500, target=0.27))))
#
runs <- mps(om, ctrl=control, args=list(iy=2021, fy=2026), hcr=list(target=seq(0.15, 0.50, length=10)))
}
