% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance}
\alias{performance}
\alias{performance,FLQuants-method}
\alias{performance,FLStocks-method}
\alias{performance,list-method}
\alias{performance,FLmse-method}
\title{Compute performance indicators}
\usage{
\S4method{performance}{FLStock}(
  x,
  indicators,
  refpts = FLPar(),
  years = as.character(seq(dims(x)$minyear, dims(x)$maxyear)),
  metrics = FLCore::metrics(x),
  probs = NULL,
  mp = NULL
)

\S4method{performance}{FLQuants}(
  x,
  indicators,
  refpts = FLPar(),
  years = setNames(list(dimnames(x[[1]])$year), nm = dims(x[[1]])$maxyear),
  probs = c(0.1, 0.25, 0.5, 0.75, 0.9),
  mp = NULL
)

\S4method{performance}{FLStocks}(
  x,
  indicators,
  refpts = FLPar(),
  years = dims(x[[1]])$maxyear,
  metrics = FLCore::metrics,
  probs = NULL,
  grid = missing,
  mp = NULL,
  mc.cores = 1
)

\S4method{performance}{list}(
  x,
  indicators,
  refpts = FLPar(),
  years = dims(x[[1]])$maxyear,
  probs = NULL,
  grid = missing,
  mp = NULL,
  mc.cores = 1
)

\S4method{performance}{FLmse}(
  x,
  indicators,
  refpts = x@refpts,
  years = as.character(seq(dims(x)$minyear, dims(x)$maxyear)),
  metrics = FLCore::metrics(stock(x)),
  probs = NULL,
  mp = NULL
)
}
\arguments{
\item{indicators}{Indicators to be computed, as formula, name and description, \code{list}}

\item{refpts}{Reference points for calculations, \code{list}}

\item{years}{Years on which indicators should be computed, defaults to last year of input FLQuants}

\item{run}{Object holding the results of forward projections, as a named
\code{FLQuants}}
}
\value{
data.table Results of computing performance indicators.
}
\description{
TODO
}
\details{
Each indicator is an object of class list object, with three elements, the
first two of them compulsory:
\itemize{
\item An unnamed element of class \emph{formula}, e.g. \code{yearMeans(SB/SB0)}.
\item name: A short name to be output on tables and plots, of class character,
e.g. "SB/SB\link{0}".
\item desc: A longer description of the indicator, of class character, e.g. "Mean
spawner biomass relative to unfished"
}

Each indicator \code{formula} is evaluated against the \emph{metrics} and \emph{refpts} used
in the function call. Formulas can thus use (i) the names of the \code{FLQuants}
object or of the object returned by the call to \code{metrics()}, (ii) of the
\emph{params} in the \emph{refpts} object and, for all classes but \code{FLQuants}, (iii)
functions that can be called on \emph{object}. See examples below for the
necessary matching between \emph{metrics}, \emph{refpts} and the indicators formulas.
}
\examples{

# LOAD example FLmse object
data(p4om)
# GENERATE run from last 16 years of OM
run <- window(stock(om), start=2000, end=2015)
# DEFINE indicators
indicators <- list(
  dCatch=list(~yearMeans(C[, -1]/C[, -dims(C)$year]),
    name="mean(C[t] / C[t-1])",
    desc="Mean absolute proportional change in catch"),
  varCatch=list(~yearVars(C),
    name="var(C)",
    desc="Variance in catch"),
  varF=list(~yearVars(F),
    name="var(F)",
    desc="Variance in fishing mortality"))
# COMPUTE performance
performance(run, indicators, refpts=FLPar(MSY=110000),
  metrics=list(C=catch, F=fbar), years=list(2000:2015))
# Minimum indicator, named list with formula and name
performance(run, indicators=list(CMSY=list(~C/MSY, name="CMSY")),
  refpts=FLPar(MSY=110000), metrics=list(C=catch, F=fbar),
  years=list(2000:2015))
# return quantiles
performance(run, indicators, refpts=FLPar(MSY=110000),
  metrics=list(C=catch, F=fbar), years=list(2000:2015),
  probs=c(0.05, 0.25, 0.50, 0.75, 0.95))
}
\seealso{
\code{\link{FLQuants}}
}
\author{
Iago Mosqueira, EC JRC
}
\keyword{utilities}
