% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance}
\alias{performance}
\alias{performance,FLStock-method}
\alias{performance,FLStocks-method}
\alias{performance,list-method}
\alias{performance,FLom-method}
\title{Compute performance statistics}
\usage{
\S4method{performance}{FLQuants}(
  x,
  statistics = mse::statistics[c("C", "F", "SB", "AAVC")],
  refpts = FLPar(),
  years = setNames(nm = dimnames(x[[1]])$year[-1]),
  probs = c(0.1, 0.25, 0.5, 0.75, 0.9),
  om = NULL,
  type = NULL,
  run = NULL,
  mp = paste(c(om, type, run), collapse = "_"),
  ...
)

\S4method{performance}{FLStock}(x, metrics = FLCore::metrics(x), ...)

\S4method{performance}{FLStocks}(
  x,
  statistics,
  refpts = FLPar(),
  years = seq(dims(x[[1]])$minyear + 1, dims(x[[1]])$maxyear),
  metrics = FLCore::metrics,
  probs = NULL,
  grid = missing,
  mc.cores = 1,
  ...
)

\S4method{performance}{list}(
  x,
  statistics,
  refpts = FLPar(),
  years = seq(dims(x[[1]])$minyear + 1, dims(x[[1]])$maxyear),
  probs = NULL,
  grid = "missing",
  mc.cores = 1,
  ...
)

\S4method{performance}{FLom}(
  x,
  refpts = x@refpts,
  statistics = mse::statistics[c("C", "F", "SB")],
  metrics = NULL,
  om = name(x),
  ...
)
}
\arguments{
\item{statistics}{statistics to be computed, as formula, name and description, \code{list}}

\item{refpts}{Reference points for calculations, \code{list}}

\item{years}{Years on which statistics should be computed, defaults to last year of input FLQuants}

\item{run}{Object holding the results of forward projections, as a named
\code{FLQuants}}
}
\value{
data.table Results of computing performance statistics.
}
\description{
TODO
}
\details{
Each statistics is an object of class list object, with three elements, the
first two of them compulsory:
\itemize{
\item An unnamed element of class \emph{formula}, e.g. \code{yearMeans(SB/SB0)}.
\item name: A short name to be output on tables and plots, of class character,
e.g. "SB/SB\link{0}".
\item desc: A longer description of the statistics, of class character, e.g. "Mean
spawner biomass relative to unfished"
}

Each statistic \code{formula} is evaluated against the \emph{metrics} and \emph{refpts} used
in the function call. Formulas can thus use (i) the names of the \code{FLQuants}
object or of the object returned by the call to \code{metrics()}, (ii) of the
\emph{params} in the \emph{refpts} object and, for all classes but \code{FLQuants}, (iii)
functions that can be called on \emph{object}. See examples below for the
necessary matching between \emph{metrics}, \emph{refpts} and the statistics formulas.
}
\examples{

# LOAD example FLmse object
data(sol274)
# GENERATE pseudo-run from last 20 years of OM
run <- window(stock(om), start=2012, end=2021)
# DEFINE statistics
statistics <- list(
  dCatch=list(~yearMeans(C[, -1]/C[, -dims(C)$year]),
    name="mean(C[t] / C[t-1])",
    desc="Mean absolute proportional change in catch"),
  varCatch=list(~yearVars(C),
    name="var(C)",
    desc="Variance in catch"),
  varF=list(~yearVars(F),
    name="var(F)",
    desc="Variance in fishing mortality"))
# COMPUTE performance
performance(run, statistics, refpts=FLPar(MSY=110000),
  metrics=list(C=catch, F=fbar), years=list(short=2016:2018, long=2016:2021))
# Minimum statistic, named list with formula and name
performance(run, statistics=list(CMSY=list(~yearMeans(C/MSY), name="CMSY")),
  refpts=FLPar(MSY=110000), metrics=list(C=catch, F=fbar),
  years=list(2012:2021))
# return quantiles
performance(run, statistics, refpts=FLPar(MSY=110000),
  metrics=list(C=catch, F=fbar), years=list(2012:2021),
  probs =  c(0.05, 0.1, 0.25, 0.5, 0.75, 0.9, 0.95))
# DEFINE statistics without summaries
statistics <- list(
  CMSY=list(~yearMeans(C/MSY),
    name="CMSY",
    desc="Catch over MSY"))
# COMPUTE performance
perf <- performance(run, statistics, refpts=FLPar(MSY=110000),
  metrics=list(C=catch), years=list(2012:2021))
# COMPUTE summaries
perf[, .(CMSY=mean(data))]
perf <- performance(FLStocks(B=run, A=run), statistics, 
  refpts=FLPar(MSY=110000), metrics=list(C=catch), years=list(2012:2015))
}
\seealso{
\code{\link{FLQuants}}
}
\author{
Iago Mosqueira, EC JRC
}
\keyword{utilities}
