% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\docType{methods}
\name{performance}
\alias{performance}
\alias{performance,FLQuants-method}
\alias{performance,FLStock-method}
\alias{performance,FLStocks-method}
\alias{performance,list-method}
\title{Compute performance indicators}
\usage{
\S4method{performance}{FLStock}(x, indicators, refpts,
  years = as.character(seq(dims(x)$minyear, dims(x)$maxyear)),
  metrics = FLCore::metrics(x), probs = NULL, mp = NULL)

\S4method{performance}{FLQuants}(x, indicators, refpts,
  years = dims(x[[1]])$maxyear, probs = c(0.1, 0.25, 0.5, 0.75, 0.9),
  mp = NULL)

\S4method{performance}{FLStocks}(x, indicators, refpts,
  years = dims(x[[1]])$maxyear, metrics = FLCore::metrics,
  probs = NULL, grid = missing, mp = NULL, mc.cores = 1)

\S4method{performance}{list}(x, indicators, refpts,
  years = dims(x[[1]])$maxyear, probs = NULL, grid = missing,
  mp = NULL, mc.cores = 1)
}
\arguments{
\item{indicators}{Indicators to be computed, as formula, name and description, \code{list}}

\item{refpts}{Reference points for calculations, \code{list}}

\item{years}{Years on which indicators should be computed, defaults to last year of input FLQuants}

\item{run}{Object holding the results of forward projections, as a named \code{FLQuants}}
}
\value{
data.frame Results of computing performance indicators
}
\description{
Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque eleifend
odio ac rutrum luctus. Aenean placerat porttitor commodo. Pellentesque eget porta
libero. Pellentesque molestie mi sed orci feugiat, non mollis enim tristique.
Suspendisse eu sapien vitae arcu lobortis ultrices vitae ac velit. Curabitur id
nunc euismod ante fringilla lobortis. Aliquam ullamcorper in diam non placerat.
}
\details{
Aliquam sagittis feugiat felis eget consequat. Praesent eleifend dolor massa,
vitae faucibus justo lacinia a. Cras sed erat et magna pharetra bibendum quis in
mi. Sed sodales mollis arcu, sit amet venenatis lorem fringilla vel. Vivamus vitae
ipsum sem. Donec malesuada purus at libero bibendum accumsan. Donec ipsum sapien,
feugiat blandit arcu in, dapibus dictum felis.
}
\examples{

data(cod)
indicators <- list(
  T1=list(~yearMeans(C[, -1]/C[, -dims(C)$year]), name="mean(C[t] / C[t-1])",
    desc="Mean absolute proportional change in catch"),
  T2=list(~yearVars(C), name="var(C)", desc="Variance in catch"),
  T3=list(~yearVars(F), name="var(F)", desc="Variance in fishing mortality"))
run <- window(cod, start=20)
performance(run, indicators, refpts=FLPar(MSY=0),
  metrics=list(C=catch, F=fbar), years=list(20:25, 20:30))
}
\seealso{
\code{\link{FLQuants}}
}
\author{
Iago Mosqueira, EC JRC
}
\keyword{utilities}
