% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance}
\alias{performance}
\title{Compute performance indicators}
\usage{
performance(x, ...)
}
\arguments{
\item{run}{Object holding the results of forward projections, as a named \code{FLQuants}}

\item{refpts}{Reference points for calculations, \code{list}}

\item{indicators}{Indicators to be computed, as formula, name and description, \code{list}}

\item{years}{Years on which indicators should be computed, defaults to last year of input FLQuants}
}
\value{
data.frame Results of computing performance indicators
}
\description{
Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque eleifend
odio ac rutrum luctus. Aenean placerat porttitor commodo. Pellentesque eget porta
libero. Pellentesque molestie mi sed orci feugiat, non mollis enim tristique. 
Suspendisse eu sapien vitae arcu lobortis ultrices vitae ac velit. Curabitur id 
nunc euismod ante fringilla lobortis. Aliquam ullamcorper in diam non placerat.
}
\details{
Aliquam sagittis feugiat felis eget consequat. Praesent eleifend dolor massa, 
vitae faucibus justo lacinia a. Cras sed erat et magna pharetra bibendum quis in 
mi. Sed sodales mollis arcu, sit amet venenatis lorem fringilla vel. Vivamus vitae 
ipsum sem. Donec malesuada purus at libero bibendum accumsan. Donec ipsum sapien, 
feugiat blandit arcu in, dapibus dictum felis.
}
\examples{

data(cod)
indicators <- list(T1=list(~yearMeans(C[, -1]/C[, -dims(C)$year]),
  name="mean(C_t / C_t-1)", desc="Mean absolute proportional change in catch"),
  T2=list(~yearVars(C), name="var(C)", desc="Variance in catch"),
  T3=list(~yearVars(F), name="var(F)", desc="Variance in fishing mortality"))
run <- window(cod, start=20)
performance(run, indicators, refpts=FLPar(MSY=0))
}
\author{
Iago Mosqueira, EC JRC
}
\seealso{
\code{\link{FLQuants}}
}
\keyword{utilities}

