% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/statistics.R
\docType{data}
\name{statistics}
\alias{statistics}
\title{A complete set of performance statistics}
\format{
A named list with elements of class list, each containing three elements: a formula, a long name ('name'), and a description ('desc'), both of type 'character':
\describe{
\item{SB}{SB: Mean spawner biomass}
\item{SB0}{SB/SB[0]: Mean spawner biomass relative to unfished}
\item{minSB0}{min(SB/SB[0]): Minimum spawner biomass relative to unfished}
\item{SBMSY}{SB/SB[MSY]: Mean spawnwer biomass relative to SBMSY}
\item{F}{F: Mean fishing mortality}
\item{Ftarget}{F/F[target]: Mean fishing mortality relative to target}
\item{FMSY}{F/F[MSY]: Mean fishing mortality relative to FMSY}
\item{green}{P(Green): Probability of being in Kobe green quadrant}
\item{orange}{P(Orange): Probability of being in Kobe orange quadrant}
\item{yellow}{P(Yellow): Probability of being in Kobe yellow quadrant}
\item{red}{P(Red): Probability of being in Kobe red quadrant}
\item{PSBMSY}{P(SB>=SB[MSY]): Probability of SB greater or equal to SBMSY}
\item{PSBlim}{P(SB>SB[limit]): Probability that spawner biomass is above SBlim}
\item{PSB20B0}{P(SB > 0.20 \%*\% SB[0]): Probability that spawner biomass is above 20\% SB\link{0}}
\item{risk1}{mean(P(SB<B[limit])): ICES Risk 1, mean probability that spawner biomass is below Blim}
\item{risk2}{once(P(SB<B[limit])): ICES Risk 2, probability that spawner biomass is above Blim once}
\item{risk3}{max(P(SB>B[limit])): ICES Risk 3, max probability that spawner biomass is above Blim}
\item{C}{mean(C): Mean catch over years}
\item{CMSY}{C/MSY: Mean proportion of MSY}
\item{AAVC}{AAV(C): Average annual variability in catch}
\item{IACC}{IAC(C): Percentage inter-annual change in catch}
\item{PC0}{P(shutdown): Probability of fishery shutdown, defined as catch less than 10\% of MSY}
}

A list containing named elements, each of which represents a specific statistic. Each of them contains:
\itemize{
\item \code{formula}: A formula defining how the metric is calculated.
\item \code{name}: A short descriptive name. This can contain \code{\link[=plotmath]{plotmath()}} expressions to be parsed by plot functions.
\item \code{desc}: A more detailed description of the metric.
}
}
\usage{
data(statistics)

statistics
}
\description{
A list containing a large number of performance statistics:

A list of key statistics used in fisheries management to evaluate the performance
of management procedures. This is a long list that contains statistics that measure
very similar outcomes, so a subset should be made of those most relevant to the
management objectives of interest.
}
\details{
Performance statistics are used by the \code{\link[=performance]{performance()}} method to compute time series,
or aggregates along time, of quantities of interest related to the result of applying
a particular management procedure to an operating model. They combione \code{\link[FLCore:metrics]{FLCore::metrics()}}
computed from the projected stocks, populations and fisheries, with biological,
economic or other reference points, but can also use the results of calculations and
decisions carried out by the MP.

Of the three elements in the list used to define each statistic, the first unnamed
element, of class 'formula' is the one evaluated by \code{\link[=performance]{performance()}}. The formula is
evaluated with access to the reference points of the OM, contained in the \code{refpts}
slot, a set of \link[FLCore:metrics]{FLCore::metrics} obtained from the projected OM, the contents of the
\link{tracking} table with decisions and outputs from the MP internal calcultions, as well
as any function available in the workspace.

The statistics currently included are:
\describe{
\item{SB}{Spawner biomass in tonnes (\eqn{SB}), from the \code{SB} metric.}
\item{SB0}{Spawner biomass relative to unfished (\eqn{SB/SB0}), requires the \code{SB0} refpt.}
\item{minSB0}{Minimum spawner biomass relative to unfished (\eqn{min(SB/SB0)}) requires the \code{SB0} refpt..}
\item{SBMSY}{Spawner biomass relative to \eqn{SB[MSY]} (\eqn{SB/SB[MSY]}) requires the \code{SBMSY} refpt..}
\item{R}{Recruitment (\eqn{R}), from the \code{R} metric.}
\item{F}{Fishing mortality (\eqn{F}), from the \code{F} metric.}
\item{Ftarget}{Fishing mortality relative to target (\eqn{F/F[target]}) requires the \code{Ftarget} refpt.}
\item{FMSY}{Fishing mortality relative to \eqn{F[MSY]} (\eqn{F/F[MSY]}) requires the \code{FMSY} refpt.}
\item{green}{Probability of being in the Kobe green quadrant (\eqn{P(Green)}),  requires the \code{SBMSY} and \code{FMSY} refpts.}
\item{orange}{Probability of being in the Kobe orange quadrant (\eqn{P(Orange)}),  requires the \code{SBMSY} and \code{FMSY} refpts.}
\item{yellow}{Probability of bein,  requires the \code{SBMSY} and \code{FMSY} refptsg in the Kobe yellow quadrant (\eqn{P(Yellow)}),  requires the \code{SBMSY} and \code{FMSY} refpts.}
#'   \item{red}{Probability of being in the Kobe red quadrant (\eqn{P(Red)}),  requires the \code{SBMSY} and \code{FMSY} refpts.}
\item{PSBMSY}{Probability that spawner biomass is greater than or equal to \eqn{SB[MSY]} (\eqn{P(SB>=SB[MSY])}),  requires the \code{SBMSY} refpt.}
\item{PSBlim}{Probability that spawner biomass is above \eqn{SB[lim]} (\eqn{P(SB>SB[lim])}),  requires the \code{SBlim} refpt.}
\item{PSB20B0}{Probability that spawner biomass is above 20\% of \eqn{SB0} (\eqn{P(SB > 0.20 \%*\% SB0)}),  requires the \code{SB0}refpt.}
\item{risk1}{ICES Risk 1: Probability that spawner biomass is below \eqn{B[lim]} (\eqn{P(SB<B[lim])}), requires the \code{SBlim} refpt.}
\item{risk2}{ICES Risk 2: Probability that spawner biomass falls below \eqn{B[lim]} at least once (\eqn{once(P(SB<B[lim]))}), requires the \code{SBlim} refpt.}
\item{risk3}{ICES Risk 3: Maximum probability that spawner biomass is below \eqn{B[lim]} (\eqn{max(P(SB<B[lim]))}), requires the \code{SBlim} refpt.}
\item{C}{Catch in tonnes (\eqn{C[t]}), from the \code{C} metric.}
\item{CMSY}{Proportion of maximum sustainable yield (\eqn{C/MSY}), requires the \code{MSY} refpt.}
\item{IACC}{Percentage inter-annual change in catch (\eqn{IAC(C)}), from the \code{C} metric.}
\item{PIACC20}{Probability that the inter-annual change in catch being less than 20\% (\eqn{P(IAC(C)<0.20)}), from the \code{C} metric.}
\item{PC0}{Probability of fishery shutdown (\eqn{P(shutdown)}), defined as catch falling below 10\% of MSY, so requires the \code{MSY} refpt.}
}
}
\examples{
data(statistics)
# Access a specific statistic
statistics$SBMSY
}
\keyword{datasets}
